Consolidates title elements to bypass new default react rendering
Previously, react was rendering interpolated values using html comments
as separators, which is fine in 99% of all elements except on the title
which requires a single text node for child. This was causing weird
title rendering issues (like missing spaces after the `-` separator) on
some browsers, as well as a warning message.

See
https://github.com/vercel/next.js/discussions/38256#discussioncomment-3070196
for more info.